misc/p11n-trophy/challenge-src/requirements.txt
pillow==10.3.0
requests==2.32.3
Flask==3.0.3
